Extract process killing logic to the DataClearing interface

+++ app/src/main/java/com/duckduckgo/app/fire/AutomaticDataClearer.kt
@@ -183,7 +183,18 @@ class AutomaticDataClearer @Inject constructor(
     }
 
     override fun onExit() {
-        clearDataAction.killProcess()
+        launch(dispatchers.io()) {
+            // the app does not have any activity in CREATED state we kill the process
+            val shouldKillProcess = if (androidBrowserConfigFeature.moreGranularDataClearingOptions().isEnabled()) {
+                dataClearing.shouldKillProcessOnExit()
+            } else {
+                settingsDataStore.automaticallyClearWhatOption != ClearWhatOption.CLEAR_NONE
+            }
+
+            if (shouldKillProcess) {
+                clearDataAction.killProcess()
+            }
+        }
     }
 
     private fun scheduleBackgroundTimerToTriggerClear(durationMillis: Long) {
